Move `FilterEngine` into dedicated module
